EnemeyMovement component update
When far enough from the player the enemy moves randomly in a certain bounded radius. But when the enemy is close enough from the player it chases the player up to a certain radius of pursuit. And at all times, even during the pursuit/chase, if the player gets far enough from the enemy, it automatically comes back to its initial position, unless it is triggered again by the player.

In addition, collision bugs were delt by making the enemy bounce whenever it collides with an obstacle.

Player/enemy collisions are still not working good. It must be corrected.