FEATURE-DisplayRedDotPosition
When the map is displayed, a red dot appears in at the player's position on the map, so the player knows where he is located in the map. At the moment the feature works in both fullscreen and not fullscreen. The feature works but there are still a few bugs that need to be taken care of.