Feature-MiniMAP
This feature allows to display the minimap in the corner. For now it doesnt work. I displayed a black rectangle in the corner of the screen and the map moves as the player moves but it is not in the correct scaling. I need to work on it but it is a hard feature.