Feature-Ping
Added a feauture for the player to be able to ping a location on the map when in fullscreen. Player can click on any location on the screen. The code is correctly implemented right now but it doesnt appear maybe because of mac, but I feel the code is correct. I am still trying to debug.